+++ tensortrade/__init__.py
@@ -1,7 +1,7 @@
 from .core import *  # core must come before all of the other imports
 
 from . import data
-from tensortrade.oms import orders, wallets, exchanges, instruments
+from tensortrade.oms import orders, wallets, instruments, exchanges, services
 from . import stochastic
 from . import agents
