+++ tensortrade/oms/services/slippage/random_slippage_model.py
@@ -14,7 +14,7 @@
 
 import numpy as np
 
-from tensortrade.oms.exchanges.services.slippage.slippage_model import SlippageModel
+from tensortrade.oms.services.slippage.slippage_model import SlippageModel
 from tensortrade.oms.orders import Trade, TradeType, TradeSide
